# feat: faster span and trace id generation