--- a/packages/opentelemetry-core/src/platform/node/id.ts
+++ b/packages/opentelemetry-core/src/platform/node/id.ts
@@ -33,8 +33,18 @@ const SHARED_BUFFER = Buffer.allocUnsafe(TRACE_ID_BYTES);
 function getIdGenerator(bytes: number): () => string {
   return function generateId() {
     for (let i = 0; i < bytes / 4; i++) {
-      SHARED_BUFFER.writeUInt32BE((Math.random() * 2 ** 32) >>> 0, i * 4);
+      SHARED_BUFFER.writeUInt32BE(Math.random() * 2 ** 32, i * 4);
     }
+
+    // If buffer is all 0, set the last byte to 1 to guarantee a valid w3c id is generated
+    for (let i = 0; i < bytes; i++) {
+      if (SHARED_BUFFER[i] > 0) {
+        break;
+      } else if (i === bytes - 1) {
+        SHARED_BUFFER[bytes - 1] = 1;
+      }
+    }
+
     return SHARED_BUFFER.toString('hex', 0, bytes);
   };
 }