Fixes #1334

/cc @anuraaga 

# Benchmarks

Noticeable performance improvements in the `BasicTracerProvider` benchmarks with small numbers of attributes.

Browser benchmarks: https://jsperf.com/opentelemetry-random-id-generation/1

## ID Generation Benchmark

```javascript
const Benchmark = require('benchmark');
const benchmarks = require('beautify-benchmark');
const crypto = require('crypto');

const suite = new Benchmark.Suite();

const SHARED_BUFFER = Buffer.allocUnsafe(16);
const bytes = 16;

suite
  .add('old id generation', function () {
    return crypto.randomBytes(bytes).toString('hex');
  })
  .add('new id generation', function () {
    for (let i = 0; i < bytes / 4; i++) {
      // unsigned right shift drops decimal part of the number
      // it is required because if a number between 2**32 and 2**32 - 1 is generated, an out of range error is thrown by writeUInt32BE
      SHARED_BUFFER.writeUInt32BE((Math.random() * 2 ** 32) >>> 0, i * 4);
    }

    // If buffer is all 0, set the last byte to 1 to guarantee a valid w3c id is generated
    for (let i = 0; i < bytes; i++) {
      if (SHARED_BUFFER[i] > 0) {
        break;
      } else if (i === bytes - 1) {
        SHARED_BUFFER[bytes - 1] = 1;
      }
    }

    return SHARED_BUFFER.toString('hex', 0, bytes);
  })
  // add listeners
  .on('cycle', function (event) {
    benchmarks.add(event.target);
  })
  .on('complete', function () {
    benchmarks.log();
    console.log('Fastest is ' + this.filter('fastest').map('name'));
  })
  .run();
```

```
  2 tests completed.

  old id generation x   411,744 ops/sec ±1.16% (90 runs sampled)
  new id generation x 4,642,384 ops/sec ±0.75% (93 runs sampled)
```

## New ID Generation

```
Beginning NoopTracerProvider Benchmark...
  5 tests completed.

  #startSpan                     x 902,557,527 ops/sec ±3.32% (20 runs sampled)
  #startSpan:parent              x 896,869,970 ops/sec ±3.00% (20 runs sampled)
  #startSpan with attribute      x 916,297,055 ops/sec ±3.43% (20 runs sampled)
  #startSpan with 30 attributes  x  11,340,599 ops/sec ±1.92% (20 runs sampled)
  #startSpan with 100 attributes x   3,670,278 ops/sec ±1.17% (20 runs sampled)

Beginning BasicTracerProvider Benchmark...
  5 tests completed.

  #startSpan                     x 340,802 ops/sec ±2.05% (20 runs sampled)
  #startSpan:parent              x 187,078 ops/sec ±1.77% (20 runs sampled)
  #startSpan with attribute      x 324,449 ops/sec ±2.44% (20 runs sampled)
  #startSpan with 30 attributes  x  76,231 ops/sec ±1.39% (20 runs sampled)
  #startSpan with 100 attributes x   5,920 ops/sec ±1.83% (20 runs sampled)

Beginning BasicTracerProvider with SimpleSpanProcessor Benchmark...
  5 tests completed.

  #startSpan                     x 247,847 ops/sec ±8.28% (20 runs sampled)
  #startSpan:parent              x 133,570 ops/sec ±1.88% (20 runs sampled)
  #startSpan with attribute      x 182,922 ops/sec ±17.35% (20 runs sampled)
  #startSpan with 30 attributes  x  52,325 ops/sec ±0.30% (20 runs sampled)
  #startSpan with 100 attributes x   5,396 ops/sec ±5.11% (20 runs sampled)

Beginning BasicTracerProvider with BatchSpanProcessor Benchmark...
  5 tests completed.

  #startSpan                     x 224,913 ops/sec ±1.11% (20 runs sampled)
  #startSpan:parent              x 117,169 ops/sec ±1.17% (20 runs sampled)
  #startSpan with attribute      x 183,762 ops/sec ±16.86% (20 runs sampled)
  #startSpan with 30 attributes  x  21,828 ops/sec ±20.00% (20 runs sampled)
  #startSpan with 100 attributes x   5,087 ops/sec ±2.23% (20 runs sampled)
```

## Old ID Generation

```
Beginning NoopTracerProvider Benchmark...
  5 tests completed.

  #startSpan                     x 876,884,335 ops/sec ±3.21% (20 runs sampled)
  #startSpan:parent              x 883,984,177 ops/sec ±3.07% (20 runs sampled)
  #startSpan with attribute      x 900,037,749 ops/sec ±4.01% (20 runs sampled)
  #startSpan with 30 attributes  x  11,137,143 ops/sec ±1.61% (20 runs sampled)
  #startSpan with 100 attributes x   3,661,880 ops/sec ±1.34% (20 runs sampled)

Beginning BasicTracerProvider Benchmark...
  5 tests completed.

  #startSpan                     x 140,603 ops/sec ±3.06% (20 runs sampled)
  #startSpan:parent              x  88,120 ops/sec ±1.48% (20 runs sampled)
  #startSpan with attribute      x 140,540 ops/sec ±2.78% (20 runs sampled)
  #startSpan with 30 attributes  x  55,164 ops/sec ±2.19% (20 runs sampled)
  #startSpan with 100 attributes x   5,852 ops/sec ±1.26% (20 runs sampled)

Beginning BasicTracerProvider with SimpleSpanProcessor Benchmark...
  5 tests completed.

  #startSpan                     x 127,693 ops/sec ±2.09% (20 runs sampled)
  #startSpan:parent              x  75,054 ops/sec ±5.69% (20 runs sampled)
  #startSpan with attribute      x 126,848 ops/sec ±1.59% (20 runs sampled)
  #startSpan with 30 attributes  x  42,302 ops/sec ±7.81% (20 runs sampled)
  #startSpan with 100 attributes x   4,387 ops/sec ±20.72% (20 runs sampled)

Beginning BasicTracerProvider with BatchSpanProcessor Benchmark...
  5 tests completed.

  #startSpan                     x 122,387 ops/sec ±1.67% (20 runs sampled)
  #startSpan:parent              x  69,848 ops/sec ±3.11% (20 runs sampled)
  #startSpan with attribute      x 112,986 ops/sec ±1.78% (20 runs sampled)
  #startSpan with 30 attributes  x  40,048 ops/sec ±1.52% (20 runs sampled)
  #startSpan with 100 attributes x   5,384 ops/sec ±1.47% (20 runs sampled)
```