@dyladan 2020-07-24
> I'm not really sure if we should sacrifice the built in crypto with native support for random bytes towards the pseudo Math.random, I would rather expect the opposite.

Why would you expect the opposite? There is nothing in the spec which requires the generated ids to be cryptographically random (nor does it even require them to be random). I could see the code simplicity angle, where using built-ins results in simpler code, but I don't think the code here is particularly complex, and on the browser in particular the new code is significantly simpler and more direct.

>With regards to benchmarks I would like to see the same benchmarks to be run on some benchmarking pages that compares many different browsers and then compare how much you can gain. But if the gain is couple % we have to think if it is really worth to drop crypto in favour of Math.random

The benchmarks collected here are on node, and you can see additional benchmarks in the linked issue, also collected with node. I would like to see tests across many browsers, but I think it is a fairly well-accepted fact that generating cryptographically secure random numbers is more expensive than random numbers without cryptographic guarantees. One additional thing to note aside from the performance improvement, is that this is also impervious to entropy depletion. If the system does not have enough entropy to generate cryptographically random numbers, it can pause.