@dyladan 2020-07-24
> the collision chance are really small and we aren't doing anything security sensitive with those ids so the performance gain is worth for me

This is not any more or less likely to collide. The only "loss" is that it may be now possible, but still likely very hard, to guess what the next generated ID will be.