@dyladan 2020-07-24
/cc @Flarna since you helped me with the current id generation