@dyladan 2020-07-24
@obecny I added chrome jsperf benchmark which does indeed show a 2x speedup, but I can remove the browser changes if you'd like.