@obecny 2020-07-24
> @obecny I added chrome jsperf benchmark to the PR description which does indeed show a 2x speedup, but I can remove the browser changes if you'd like.

why would you remove the browser if the speed is 2x faster ?