@dyladan 2020-07-24
> webkit browser are 2 faster, cannot check IE, now but even though it is much better

Just had a coworker test on his windows machine. The old code does not even work on IE as the ArrayBuffer does not support `slice`. New code works fine.