@obecny 2020-07-24
I added version 2 with msCrypto but it needs some polyfills I guess