@dyladan 2020-07-28
On my machine there appears to be no measurable difference between not checking, setting the last bit to 1, or looping until a non-zero byte is found.

```javascript
const Benchmark = require('benchmark');
const benchmarks = require('beautify-benchmark');
const suite = new Benchmark.Suite();

const SHARED_BUFFER = Buffer.allocUnsafe(16);

suite
  .add('without check', function () {
    for (let i = 0; i < 4; i++) {
      SHARED_BUFFER.writeUInt32BE((Math.random() * 2**32), i * 4)
    }
    SHARED_BUFFER.toString('hex', 0, 16);
  })
  .add('last bit always 1', function () {
    for (let i = 0; i < 4; i++) {
      SHARED_BUFFER.writeUInt32BE((Math.random() * 2**32), i * 4)
    }
    SHARED_BUFFER[15] = SHARED_BUFFER[15] | 1;
    SHARED_BUFFER.toString('hex', 0, 16);
  })
  .add('loop check', function () {
    for (let i = 0; i < 4; i++) {
      SHARED_BUFFER.writeUInt32BE((Math.random() * 2**32), i * 4)
    }

    for (let i = 0; i < 16; i++) {
      if (SHARED_BUFFER[i] > 0) {
        break;
      } else if (i === 15) {
        SHARED_BUFFER[15] = 1;
      }
    }

    SHARED_BUFFER.toString('hex', 0, 16);
  })
  // add listeners
  .on('cycle', function (event) {
    benchmarks.add(event.target);
  })
  .on('complete', function () {
    benchmarks.log();
    console.log('Fastest is ' + this.filter('fastest').map('name'));
  })
  .run();
```

```
  3 tests completed.

  without check     x 4,816,246 ops/sec ±0.89% (93 runs sampled)
  last bit always 1 x 4,790,933 ops/sec ±0.89% (94 runs sampled)
  loop check        x 4,776,446 ops/sec ±0.77% (95 runs sampled)

Fastest is without check,last bit always 1
```