@dyladan 2020-07-29
@open-telemetry/javascript-approvers this is done being updated I think, and I would appreciate reviews so we can get this merged.