@dyladan 2020-08-13
@open-telemetry/javascript-approvers would really appreciate some ✅ here :)